Moved the SafetyAreaManager class into the cpp
